==> src/00_data_wrangle/04_scrape_images.ipynb <==
import pandas as pd
from bs4 import BeautifulSoup
import requests
import time
from selenium import webdriver
import numpy as np

## Pull Each Team's Roster URL

team_url = 'https://web.archive.org/web/20181202093151/http://www.espn.com/nfl/players'
request = requests.get(team_url)
soup = BeautifulSoup(request.text, 'html.parser')

["https://web.archive.org" + x["href"] for x in soup.findAll('a') if ("team/" in x["href"]) and ("roster/" in x["href"])]

teams = ["https://web.archive.org" + x["href"] for x in soup.findAll('a') if ("team/" in x["href"]) and ("roster/" in x["href"])]

[x.text for x in soup.findAll('div', class_ = "mod-header stathead")]

# Getting Player Ids

teams[0].split("/name/")[1].split("/")[0].upper()

playerDf_tot = pd.DataFrame()

for team_url in teams:
    
    print("Pulling Players From " + team_url.split("/name/")[1].split("/")[0].upper())

    request = requests.get(team_url)
    soup = BeautifulSoup(request.text, 'html.parser')

    player_urls = [x for x in [x['href'] for x in soup.findAll('a') if 'href' in str(x)] if ("player/" in x)]

    player_ids = [x.split("/id/")[1] for x in player_urls]

    player_image_urls = ["https://a.espncdn.com/combiner/i?img=/i/headshots/nfl/players/full/" + 
                        x + ".png&w=350&h=254" for x in player_ids]

    player_names = np.array([x.text for x in soup.findAll('a') if 'href' in str(x)])[np.array([("player/" in x) for x in [x['href'] for x in soup.findAll('a') if 'href' in str(x)]])]

    player_positions = [x.text for x in soup.findAll('td', class_ = 'Table2__td')][2::8]

    player_schools = [x.text for x in soup.findAll('td', class_ = 'Table2__td')][7::8]

    playerDf = pd.DataFrame({"displayName": player_names,
                  "team": team_url.split("/name/")[1].split("/")[0].upper(),
                  "position": player_positions,
                  "collegeName": player_schools,
                  "playerId":player_ids,
                  "playerImageUrl":player_image_urls})
    
    playerDf_tot = pd.concat([playerDf_tot,
                              playerDf])

56*32

playerDf_tot

playerDf_tot.to_csv("/Users/James/Desktop/CoverageNet/src/00_data_wrangle/helper_tables/playerImages.csv",
                 index = False)

# Scraping College Images

team_url = 'https://www.espn.com/college-football/teams'
request = requests.get(team_url)
soup = BeautifulSoup(request.text, 'html.parser')

college_full_names = [x["title"] for x in soup.findAll('img', class_ = "Image Logo Logo__lg")]

college_team_links = ['https://espn.com' + x.find('a')["href"] for x in soup.findAll('div', class_ = "TeamLinks__Links")]

college_team_ids = [x.find('a')["href"].split('id/')[1] for x in soup.findAll('div', class_ = "TeamLinks__Links")]

college_team_images = ["https://a.espncdn.com/combiner/i?img=/i/teamlogos/ncaa/500/" +\
                        x + ".png&h=80&w=80&scale=crop&cquality=40&location=origin" for x in college_team_ids]

# gettng college name
college_name = []
for link in college_team_links:
    team_url = link
    request = requests.get(team_url)
    soup = BeautifulSoup(request.text, 'html.parser')
    name = soup.find('span', class_ = "db pr3 nowrap fw-bold").text
    college_name.append(name)
    print(name)

collegeTeamsDf = pd.DataFrame({"collegeFullName": college_full_names,
                              "collegeName": college_name,
                              "collegeId": college_team_ids,
                              "collegeImageUrl": college_team_images})

collegeTeamsDf

collegeTeamsDf.to_csv("/Users/James/Desktop/CoverageNet/src/00_data_wrangle/helper_tables/dashboard_college_info.csv",
                 index = False)
